requirements.txt
flask
flask-cors
python-dotenv
openai
anthropic
google-genai
requests
gunicorn
python-pptx
